In Cordoba report, when you import a new points file, can you make sure that the file cannot be imported if any customer’s final balance is negative after import process.
In Cordoba report, when you import a new points file, can you make sure that the file cannot be imported if any customer’s final balance is negative after import process.

e.g. if the customer’s current point is 99 and the import amount is -105, after import the final balance is -6. In this case, do not import whole file and show the error with customer name.